+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- The method `Geometry.reflected` can be used to create a reflected copy of any geometry off a plane.
+
 ## [2.8.3] - 2025-04-24
 
 ### Added